no. 1 bug fix and betty

+++ file_io/1-create_file.c
@@ -22,7 +22,7 @@ int create_file(const char *filename, char *text_content)
 	if (text_content != NULL)
 	{
 		bytes_written = write(fd, text_content, strlen(text_content));
-		if (bytes_written == -1 || bytes_written != strlen(text_content))
+		if (bytes_written == -1 || bytes_written != (int)strlen(text_content))
 		{
 			close(fd);
 			return (-1);
